refactor setup_test
