+++ internal/resources/acceptance_tests/security/app_services_security_acceptance_test.go
@@ -1,0 +1,216 @@
+package security
+
+import (
+	"fmt"
+	"os"
+	"regexp"
+	"testing"
+
+	acctest "github.com/couchbasecloud/terraform-provider-couchbase-capella/internal/testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+)
+
+func TestAppServiceResourceNoAuth(t *testing.T) {
+
+	tempId := os.Getenv("TF_VAR_auth_token")
+	os.Setenv("TF_VAR_auth_token", "")
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.TestAccPreCheck(t) },
+		ProtoV6ProviderFactories: acctest.TestAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			// Create and Read testing
+			{
+				Config:      testAccAppServiceResourceConfig(acctest.Cfg),
+				ExpectError: regexp.MustCompile("Missing Capella Authentication Token"),
+			},
+		},
+	})
+	os.Setenv("TF_VAR_auth_token", tempId)
+}
+
+func TestAppServiceResourceOrgOwner(t *testing.T) {
+
+	tempId := os.Getenv("TF_VAR_auth_token")
+	testAccCreateOrgAPI("organizationOwner")
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.TestAccPreCheck(t) },
+		ProtoV6ProviderFactories: acctest.TestAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			// Create and Read testing
+			{
+				Config: testAccAppServiceResourceConfig(acctest.Cfg),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("capella_app_service.new_app_service", "name", "test-terraform-app-service"),
+					resource.TestCheckResourceAttr("capella_app_service.new_app_service", "description", "description"),
+					resource.TestCheckResourceAttr("capella_app_service.new_app_service", "compute.cpu", "2"),
+					resource.TestCheckResourceAttr("capella_app_service.new_app_service", "compute.ram", "4"),
+					resource.TestCheckResourceAttr("capella_app_service.new_app_service", "nodes", "2"),
+				),
+			},
+		},
+	})
+	os.Setenv("TF_VAR_auth_token", tempId)
+}
+
+func TestAppServiceResourceOrgMember(t *testing.T) {
+
+	tempId := os.Getenv("TF_VAR_auth_token")
+	testAccCreateOrgAPI("organizationMember")
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.TestAccPreCheck(t) },
+		ProtoV6ProviderFactories: acctest.TestAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			// Create and Read testing
+			{
+				Config:      testAccAppServiceResourceConfig(acctest.Cfg),
+				ExpectError: regexp.MustCompile("Access Denied"),
+			},
+		},
+	})
+	os.Setenv("TF_VAR_auth_token", tempId)
+}
+
+func TestAppServiceResourceProjCreator(t *testing.T) {
+
+	tempId := os.Getenv("TF_VAR_auth_token")
+	testAccCreateOrgAPI("projectCreator")
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.TestAccPreCheck(t) },
+		ProtoV6ProviderFactories: acctest.TestAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			// Create and Read testing
+			{
+				Config:      testAccAppServiceResourceConfig(acctest.Cfg),
+				ExpectError: regexp.MustCompile("Access Denied"),
+			},
+		},
+	})
+	os.Setenv("TF_VAR_auth_token", tempId)
+}
+
+func TestAppServiceResourceProjOwner(t *testing.T) {
+
+	tempId := os.Getenv("TF_VAR_auth_token")
+	projId := os.Getenv("TF_VAR_project_id")
+	testAccCreateProjAPI("projectCreator", projId, "projectOwner")
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.TestAccPreCheck(t) },
+		ProtoV6ProviderFactories: acctest.TestAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			// Create and Read testing
+			{
+				Config: testAccAppServiceResourceConfig(acctest.Cfg),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("capella_app_service.new_app_service", "name", "test-terraform-app-service"),
+					resource.TestCheckResourceAttr("capella_app_service.new_app_service", "description", "description"),
+					resource.TestCheckResourceAttr("capella_app_service.new_app_service", "compute.cpu", "2"),
+					resource.TestCheckResourceAttr("capella_app_service.new_app_service", "compute.ram", "4"),
+					resource.TestCheckResourceAttr("capella_app_service.new_app_service", "nodes", "2"),
+				),
+			},
+		},
+	})
+	os.Setenv("TF_VAR_auth_token", tempId)
+}
+
+func TestAppServiceResourceProjManager(t *testing.T) {
+
+	tempId := os.Getenv("TF_VAR_auth_token")
+	projId := os.Getenv("TF_VAR_project_id")
+	testAccCreateProjAPI("projectCreator", projId, "projectManager")
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.TestAccPreCheck(t) },
+		ProtoV6ProviderFactories: acctest.TestAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			// Create and Read testing
+			{
+				Config: testAccAppServiceResourceConfig(acctest.Cfg),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("capella_app_service.new_app_service", "name", "test-terraform-app-service"),
+					resource.TestCheckResourceAttr("capella_app_service.new_app_service", "description", "description"),
+					resource.TestCheckResourceAttr("capella_app_service.new_app_service", "compute.cpu", "2"),
+					resource.TestCheckResourceAttr("capella_app_service.new_app_service", "compute.ram", "4"),
+					resource.TestCheckResourceAttr("capella_app_service.new_app_service", "nodes", "2"),
+				),
+			},
+		},
+	})
+	os.Setenv("TF_VAR_auth_token", tempId)
+}
+
+func TestAppServiceResourceProjViewer(t *testing.T) {
+
+	tempId := os.Getenv("TF_VAR_auth_token")
+	projId := os.Getenv("TF_VAR_project_id")
+	testAccCreateProjAPI("projectCreator", projId, "projectViewer")
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.TestAccPreCheck(t) },
+		ProtoV6ProviderFactories: acctest.TestAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			// Create and Read testing
+			{
+				Config:      testAccAppServiceResourceConfig(acctest.Cfg),
+				ExpectError: regexp.MustCompile("Access Denied"),
+			},
+		},
+	})
+	os.Setenv("TF_VAR_auth_token", tempId)
+}
+
+func TestAppServiceResourceDatabaseDataReaderWriter(t *testing.T) {
+
+	tempId := os.Getenv("TF_VAR_auth_token")
+	projId := os.Getenv("TF_VAR_project_id")
+	testAccCreateProjAPI("projectCreator", projId, "projectDataReaderWriter")
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.TestAccPreCheck(t) },
+		ProtoV6ProviderFactories: acctest.TestAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			// Create and Read testing
+			{
+				Config:      testAccAppServiceResourceConfig(acctest.Cfg),
+				ExpectError: regexp.MustCompile("Access Denied"),
+			},
+		},
+	})
+	os.Setenv("TF_VAR_auth_token", tempId)
+}
+
+func TestAppServiceResourceDatabaseDataReader(t *testing.T) {
+
+	tempId := os.Getenv("TF_VAR_auth_token")
+	projId := os.Getenv("TF_VAR_project_id")
+	testAccCreateProjAPI("projectCreator", projId, "projectDataReader")
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.TestAccPreCheck(t) },
+		ProtoV6ProviderFactories: acctest.TestAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			// Create and Read testing
+			{
+				Config:      testAccAppServiceResourceConfig(acctest.Cfg),
+				ExpectError: regexp.MustCompile("Access Denied"),
+			},
+		},
+	})
+	os.Setenv("TF_VAR_auth_token", tempId)
+}
+
+func testAccAppServiceResourceConfig(cfg string) string {
+	return fmt.Sprintf(`
+%[1]s
+
+resource "capella_app_service" "new_app_service" {
+  organization_id = var.organization_id
+  project_id      = var.project_id
+  cluster_id      = var.cluster_id
+  name            = "test-terraform-app-service"
+  description     = "description"
+  compute = {
+    cpu = 2
+    ram = 4
+  }
+  nodes = 2
+}
+`, cfg)
+}
