+++ internal/resources/acceptance_tests/security/certificates_security_acceptance_test.go
@@ -1,4 +1,4 @@
-package security_acceptance_tests
+package security
 
 import (
 	"fmt"
