+++ internal/resources/acceptance_tests/security/organizations_security_acceptance_test.go
@@ -1,4 +1,4 @@
-package security_acceptance_tests
+package security
 
 import (
 	"encoding/json"
